requirements.txt
##### Requirements for JAX-RL #####
numpy<1.24
jax
dm-haiku
optax
gym
